HUD
HUD:
Las balas se muestran en la parte inferior izquierda
Hay un CrossHair en el medio de la pantalla

TMPro:
Se incluyo TMPro en el proyecto

+++ 3dshooter/Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    void ApplyDamage(float damageTaken);
+    void ApplyDamage(int damageTaken);
 }
